truy cap voi JDBC

+++ src/main/java/ViewServlet.java
@@ -47,7 +47,7 @@ public class ViewServlet extends HttpServlet {
                     result += "<td>" +rsUsers.getString(4) + "</td>";
                     result += "<td>" +rsUsers.getString(5) + "</td>";
                     result += "<td><a href='EditServlet?id=" + rsUsers.getInt(1) + "'>Edit</a></td>";
-                    result += "<td><a href='DeleteServlet?id=" + rsUsers.getInt(1) + "'>Delete</a></td>";
+                    result += "<td><a href='DeleteServlet?id=" + rsUsers.getInt(1) + "' onclick='return confirm(\"Are you sure you want to delete this user?\");'>Delete</a></td>";
                     result += "</tr>";
                 }
                 out.println(result);
